Move dataplane/ to pkg/
