+++ pkg/dataplane/grpc/grpc_dataplane.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/containernetworking/cni/pkg/skel"
 	"github.com/containernetworking/cni/pkg/types/current"
-	"github.com/projectcalico/cni-plugin/dataplane/grpc/proto"
+	"github.com/projectcalico/cni-plugin/pkg/dataplane/grpc/proto"
 	"github.com/projectcalico/cni-plugin/pkg/types"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
