+++ pkg/dataplane/grpc/proto/doc.go
@@ -4,6 +4,6 @@ package proto
 
 // The proto package defines a simple interface between the CNI plugin
 // and a "dataplane driver", which can be used to interface the Calico
-// CNI plugin with a custom dataplane, much like felix's dtaplane
+// CNI plugin with a custom dataplane, much like felix's dataplane
 // abstraction allows to use Calico's policy engine with a non-kernel
 // dataplane
